test: refactor react for role tests to new approach
